+++ packages/collections-orm/src/QueryPageable.php
@@ -40,6 +40,7 @@ class QueryPageable implements PageableInterface
      */
     public function __construct(
         private QueryBuilder $queryBuilder,
+        private ?string $indexBy = null,
         private readonly int $itemsPerPage = 50,
         private readonly CountStrategy $countStrategy = CountStrategy::Restrict,
         private ?int &$count = null,