+++ packages/collections-orm/src/Trait/QueryBuilderTrait.php
@@ -43,6 +43,7 @@ trait QueryBuilderTrait
 
         $adapter = new QueryBuilderAdapter(
             queryBuilder: $this->queryBuilder,
+            indexBy: $this->indexBy,
         );
 
         $count = match ($this->countStrategy) {